\name{availableMethods}
\alias{availableMethods}
\title{Function that returns list of available methods.}
\usage{
  availableMethods(A, B)
}
\arguments{
  \item{A}{Dummy parameter.}

  \item{B}{Dummy parameter.}
}
\value{
  List of available methods.
}
\description{
  Function that returns list of available methods.
}
\examples{
availableMethods()
}

