\name{csd}
\alias{Sensor}
\alias{Sensor-class}
\alias{csd}
\title{Method csd.}
\usage{
  Sensor(...)
}
\arguments{
  \item{...}{constructor parameters.}
}
\description{
  Method csd.

  Class \code{Sensor} defines a gas sensor device.

  Constructor method of Sensor Class.

  Wrapper function Sensor.
}

