\name{Sensor-class}
\alias{Sensor}
\alias{Sensor-class}
\alias{csd}
\alias{csd,Sensor-method}
\alias{dsd}
\alias{dsd,Sensor-method}
\alias{ssd}
\alias{ssd,Sensor-method}
\title{Method csd.}
\usage{
  Sensor(...)
}
\arguments{
  \item{...}{constructor parameters.}
}
\description{
  Method csd.

  Method ssd.

  Method dsd.

  Class \code{Sensor} defines a gas sensor device.

  Constructor method of Sensor Class.

  Wrapper function Sensor.
}

